refactor(editor): extract small panels into modules
Mechanical module split, step 2/4: Views, Overview, Favorites +
Light-Favorites, Room-Pins and Weather-Sensors panels move to
src/editor/panels/, plus the shared entity-list drag & drop handlers.
Panels are plain render functions over the StrategyEditorHost contract.

Only deliberate deviations from verbatim: dynamic lookups on moved
lines hardened per CLAUDE.md Codacy pitfalls (Reflect.get for the
device-class table, stateFor() for hass.states, explicit field
branches in updateWeatherSensor), and the dead select-based
_addFavoriteFromSelect/_addRoomPinFromSelect paths dropped (the search
picker replaced them; no references).
